+++ src/main/java/com/epam/ta/reportportal/core/project/impl/DeleteProjectHandler.java
@@ -24,6 +24,7 @@ package com.epam.ta.reportportal.core.project.impl;
 import static com.epam.ta.reportportal.commons.Predicates.*;
 import static com.epam.ta.reportportal.commons.validation.BusinessRule.expect;
 import static com.epam.ta.reportportal.ws.model.ErrorType.PROJECT_NOT_FOUND;
+import static java.util.Arrays.asList;
 import static java.util.Collections.singletonList;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -39,6 +40,8 @@ import com.epam.ta.reportportal.exception.ReportPortalException;
 import com.epam.ta.reportportal.triggers.CascadeDeleteProjectsService;
 import com.epam.ta.reportportal.ws.model.ErrorType;
 import com.epam.ta.reportportal.ws.model.OperationCompletionRS;
+
+import java.util.Arrays;
 
 /**
  * Initial implementation of
@@ -67,8 +70,8 @@ public class DeleteProjectHandler implements IDeleteProjectHandler {
 
 		Project project = projectRepository.findOne(projectName);
 		expect(project, notNull()).verify(PROJECT_NOT_FOUND, projectName);
-		expect(project.getConfiguration().getEntryType(), not(equalTo(EntryType.PERSONAL)))
-				.verify(ErrorType.UNABLE_TO_DELETE_PERSONAL_PROJECT);
+		expect(project.getConfiguration().getEntryType(), and(asList(not(equalTo(EntryType.PERSONAL)), not(equalTo(EntryType.UPSA)))))
+				.verify(ErrorType.PROJECT_UPDATE_NOT_ALLOWED, project.getConfiguration().getEntryType());
 		Iterable<ExternalSystem> externalSystems = externalSystemRepository.findAll(project.getConfiguration().getExternalSystem());
 
 		try {